Qualifications

+++ src/main/java/edu/eci/ieti/uniwheels/model/Calificacion.java
@@ -5,13 +5,13 @@ import org.springframework.data.annotation.Id;
 public class Calificacion {
     @Id
     public String id;
-    public int valor;
+    public double valor;
 
 
     public Calificacion(){
 
     }
-    public Calificacion( int valor) {
+    public Calificacion( double valor) {
         this.valor = valor;
     }
 
@@ -23,11 +23,11 @@ public class Calificacion {
         this.id = id;
     }
 
-    public int getValor() {
+    public double getValor() {
         return valor;
     }
 
-    public void setValor(int valor) {
+    public void setValor(float valor) {
         this.valor = valor;
     }
 }
